# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Property Management API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Property {
  id: number;
  name: string;
  address: string;
  /** @nullable */
  city?: string | null;
  /** @nullable */
  postcode?: string | null;
  /** @nullable */
  notes?: string | null;
  totalRooms: number;
  occupiedRooms: number;
  expectedRent: number;
  createdAt: string;
}

export type RoomStatus = typeof RoomStatus[keyof typeof RoomStatus];


export const RoomStatus = {
  vacant: 'vacant',
  occupied: 'occupied',
} as const;

export interface Room {
  id: number;
  propertyId: number;
  name: string;
  keyTag: string;
  status: RoomStatus;
  monthlyRent: number;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  tenantId?: number | null;
  /** @nullable */
  tenantName?: string | null;
  createdAt: string;
}

export interface PropertyDetail {
  id: number;
  name: string;
  address: string;
  /** @nullable */
  city?: string | null;
  /** @nullable */
  postcode?: string | null;
  /** @nullable */
  notes?: string | null;
  totalRooms: number;
  occupiedRooms: number;
  expectedRent: number;
  createdAt: string;
  rooms: Room[];
}

export interface PropertyInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  address: string;
  city?: string;
  postcode?: string;
  notes?: string;
}

export interface PropertyUpdate {
  /** @minLength 1 */
  name?: string;
  address?: string;
  city?: string;
  postcode?: string;
  notes?: string;
}

export interface RoomInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  keyTag: string;
  /** @minimum 0 */
  monthlyRent: number;
  notes?: string;
}

export type RoomUpdateStatus = typeof RoomUpdateStatus[keyof typeof RoomUpdateStatus];


export const RoomUpdateStatus = {
  vacant: 'vacant',
  occupied: 'occupied',
} as const;

export interface RoomUpdate {
  name?: string;
  keyTag?: string;
  /** @minimum 0 */
  monthlyRent?: number;
  status?: RoomUpdateStatus;
  notes?: string;
}

export type TenantStatus = typeof TenantStatus[keyof typeof TenantStatus];


export const TenantStatus = {
  active: 'active',
  inactive: 'inactive',
} as const;

export interface Tenant {
  id: number;
  fullName: string;
  email: string;
  /** @nullable */
  phone: string | null;
  roomId: number;
  propertyName: string;
  roomName: string;
  monthlyRent: number;
  leaseStart: string;
  /** @nullable */
  leaseEnd?: string | null;
  status: TenantStatus;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}

export type TenantDetailStatus = typeof TenantDetailStatus[keyof typeof TenantDetailStatus];


export const TenantDetailStatus = {
  active: 'active',
  inactive: 'inactive',
} as const;

export type PaymentStatus = typeof PaymentStatus[keyof typeof PaymentStatus];


export const PaymentStatus = {
  paid: 'paid',
  partial: 'partial',
  unpaid: 'unpaid',
} as const;

export interface Payment {
  id: number;
  tenantId: number;
  tenantName: string;
  roomId: number;
  roomName: string;
  amount: number;
  expectedAmount: number;
  month: string;
  status: PaymentStatus;
  /** @nullable */
  paidAt: string | null;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}

export interface Document {
  id: number;
  tenantId: number;
  tenantName: string;
  fileName: string;
  fileType: string;
  /** @nullable */
  filePath?: string | null;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}

export interface TenantDetail {
  id: number;
  fullName: string;
  email: string;
  /** @nullable */
  phone: string | null;
  roomId: number;
  propertyName: string;
  roomName: string;
  monthlyRent: number;
  leaseStart: string;
  /** @nullable */
  leaseEnd?: string | null;
  status: TenantDetailStatus;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
  payments: Payment[];
  documents: Document[];
}

export interface TenantInput {
  /** @minLength 1 */
  fullName: string;
  email: string;
  phone?: string;
  roomId: number;
  leaseStart: string;
  leaseEnd?: string;
  notes?: string;
}

export type TenantUpdateStatus = typeof TenantUpdateStatus[keyof typeof TenantUpdateStatus];


export const TenantUpdateStatus = {
  active: 'active',
  inactive: 'inactive',
} as const;

export interface TenantUpdate {
  fullName?: string;
  email?: string;
  phone?: string;
  roomId?: number;
  leaseStart?: string;
  leaseEnd?: string;
  status?: TenantUpdateStatus;
  notes?: string;
}

export type PaymentInputStatus = typeof PaymentInputStatus[keyof typeof PaymentInputStatus];


export const PaymentInputStatus = {
  paid: 'paid',
  partial: 'partial',
  unpaid: 'unpaid',
} as const;

export interface PaymentInput {
  tenantId: number;
  /** @minimum 0 */
  amount: number;
  /** @minimum 0 */
  expectedAmount: number;
  month: string;
  status: PaymentInputStatus;
  paidAt?: string;
  notes?: string;
}

export type PaymentUpdateStatus = typeof PaymentUpdateStatus[keyof typeof PaymentUpdateStatus];


export const PaymentUpdateStatus = {
  paid: 'paid',
  partial: 'partial',
  unpaid: 'unpaid',
} as const;

export interface PaymentUpdate {
  /** @minimum 0 */
  amount?: number;
  /** @minimum 0 */
  expectedAmount?: number;
  month?: string;
  status?: PaymentUpdateStatus;
  paidAt?: string;
  notes?: string;
}

export interface DocumentInput {
  tenantId: number;
  /** @minLength 1 */
  fileName: string;
  fileType: string;
  filePath?: string;
  notes?: string;
}

export interface DashboardSummary {
  totalProperties: number;
  totalRooms: number;
  occupiedRooms: number;
  vacantRooms: number;
  occupancyRate: number;
  totalTenants: number;
  expectedMonthlyRent: number;
  collectedThisMonth: number;
  arrearsTotal: number;
  collectionRate: number;
}

export interface OccupancyItem {
  propertyId: number;
  propertyName: string;
  totalRooms: number;
  occupiedRooms: number;
  occupancyRate: number;
}

export interface ArrearsItem {
  tenantId: number;
  tenantName: string;
  roomName: string;
  propertyName: string;
  monthlyRent: number;
  totalPaid: number;
  arrears: number;
  monthsUnpaid: number;
}

export interface RevenueTrendItem {
  month: string;
  collected: number;
  expected: number;
}

export interface SearchResults {
  properties: Property[];
  rooms: Room[];
  tenants: Tenant[];
}

export type ListPaymentsParams = {
/**
 * @nullable
 */
tenantId?: number | null;
/**
 * @nullable
 */
roomId?: number | null;
};

export type ListDocumentsParams = {
/**
 * @nullable
 */
tenantId?: number | null;
};

export type GlobalSearchParams = {
q: string;
};