# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Property Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /tenants:
    get:
      operationId: listTenants
      tags: [tenants]
      summary: List all tenants
      responses:
        "200":
          description: List of tenants
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Tenant"
    post:
      operationId: createTenant
      tags: [tenants]
      summary: Create a tenant
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TenantInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tenant"
  /tenants/{id}:
    get:
      operationId: getTenant
      tags: [tenants]
      summary: Get a tenant
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Tenant
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TenantDetail"
        "404":
          description: Not found
    patch:
      operationId: updateTenant
      tags: [tenants]
      summary: Update a tenant
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TenantUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tenant"
        "404":
          description: Not found
    delete:
      operationId: deleteTenant
      tags: [tenants]
      summary: Delete a tenant
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found

  # ── Payments ─────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Tenant:
      type: object
      required: [id, fullName, email, phone, roomId, propertyName, roomName, monthlyRent, leaseStart, status, createdAt]
      properties:
        id:           { type: number }
        fullName:     { type: string }
        email:        { type: string }
        phone:        { type: ["string", "null"] }
        roomId:       { type: number }
        propertyName: { type: string }
        roomName:     { type: string }
        monthlyRent:  { type: number }
        leaseStart:   { type: string }
        leaseEnd:     { type: ["string", "null"] }
        status:       { type: string, enum: [active, inactive] }
        notes:        { type: ["string", "null"] }
        createdAt:    { type: string }
    TenantDetail:
      type: object
      required: [id, fullName, email, phone, roomId, propertyName, roomName, monthlyRent, leaseStart, status, createdAt, payments, documents]
      properties:
        id:           { type: number }
        fullName:     { type: string }
        email:        { type: string }
        phone:        { type: ["string", "null"] }
        roomId:       { type: number }
        propertyName: { type: string }
        roomName:     { type: string }
        monthlyRent:  { type: number }
        leaseStart:   { type: string }
        leaseEnd:     { type: ["string", "null"] }
        status:       { type: string, enum: [active, inactive] }
        notes:        { type: ["string", "null"] }
        createdAt:    { type: string }
        payments:
          type: array
          items:
            $ref: "#/components/schemas/Payment"
        documents:
          type: array
          items:
            $ref: "#/components/schemas/Document"
    TenantInput:
      type: object
      required: [fullName, email, roomId, leaseStart]
      properties:
        fullName:   { type: string, minLength: 1 }
        email:      { type: string }
        phone:      { type: string }
        roomId:     { type: number }
        leaseStart: { type: string }
        leaseEnd:   { type: string }
        notes:      { type: string }
    TenantUpdate:
      type: object
      properties:
        fullName:   { type: string }
        email:      { type: string }
        phone:      { type: string }
        roomId:     { type: number }
        leaseStart: { type: string }
        leaseEnd:   { type: string }
        status:     { type: string, enum: [active, inactive] }
        notes:      { type: string }

    # ── Payment ──────────────────────────────────────────────────────────────
    Payment:
      type: object
      required: [id, tenantId, tenantName, roomId, roomName, amount, expectedAmount, month, status, paidAt, createdAt]
      properties:
        id:             { type: number }
        tenantId:       { type: number }
        tenantName:     { type: string }
        roomId:         { type: number }
        roomName:       { type: string }
        amount:         { type: number }
        expectedAmount: { type: number }
        month:          { type: string }
        status:         { type: string, enum: [paid, partial, unpaid] }
        paidAt:         { type: ["string", "null"] }
        notes:          { type: ["string", "null"] }
        createdAt:      { type: string }
...
    Document:
      type: object
      required: [id, tenantId, tenantName, fileName, fileType, createdAt]
      properties:
        id:         { type: number }
        tenantId:   { type: number }
        tenantName: { type: string }
        fileName:   { type: string }
        fileType:   { type: string }
        filePath:   { type: ["string", "null"] }
        notes:      { type: ["string", "null"] }
        createdAt:  { type: string }
...