# lib/api-zod/src/generated/types/tenant.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Property Management API
 * OpenAPI spec version: 0.1.0
 */
import type { TenantStatus } from './tenantStatus';

export interface Tenant {
  id: number;
  fullName: string;
  email: string;
  /** @nullable */
  phone: string | null;
  roomId: number;
  propertyName: string;
  roomName: string;
  monthlyRent: number;
  leaseStart: string;
  /** @nullable */
  leaseEnd?: string | null;
  status: TenantStatus;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}