# lib/api-zod/src/generated/types/tenantDetail.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Property Management API
 * OpenAPI spec version: 0.1.0
 */
import type { Document } from './document';
import type { Payment } from './payment';
import type { TenantDetailStatus } from './tenantDetailStatus';

export interface TenantDetail {
  id: number;
  fullName: string;
  email: string;
  /** @nullable */
  phone: string | null;
  roomId: number;
  propertyName: string;
  roomName: string;
  monthlyRent: number;
  leaseStart: string;
  /** @nullable */
  leaseEnd?: string | null;
  status: TenantDetailStatus;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
  payments: Payment[];
  documents: Document[];
}